requirements.txt
agno==1.7.9
annotated-types==0.7.0
anyio==4.10.0
certifi==2025.8.3
click==8.2.1
distro==1.9.0
docstring_parser==0.17.0
dotenv==0.9.9
duckduckgo_search==8.1.1
gitdb==4.0.12
GitPython==3.1.45
groq==0.31.0
h11==0.16.0
httpcore==1.0.9
httpx==0.28.1
idna==3.10
lxml==6.0.0
markdown-it-py==4.0.0
mdurl==0.1.2
phidata==2.7.10
primp==0.15.0
pydantic==2.11.7
pydantic-settings==2.10.1
pydantic_core==2.33.2
Pygments==2.19.2
python-dotenv==1.1.1
python-multipart==0.0.20
PyYAML==6.0.2
rich==14.1.0
shellingham==1.5.4
smmap==5.0.2
sniffio==1.3.1
tomli==2.2.1
typer==0.16.0
typing-inspection==0.4.1
typing_extensions==4.14.1
